fix(doctor): preserve trust-aware tool fallback
